Don't create raw clients just for the admin bus.
This was probably here before because the impression was that otherwise the admin bus would share pub/sub clients with the app.  But `prepareAdapter` calls the main socket.io-redis function once for the app and once for the admin bus, and each call creates its own clients, so this is all good.