Upgrade dev dep of socket.io-redis to 3.1.0, mute warnings from adapter
We already handle "end" and "ready" from the individual pub/sub clients, and mute their "error" events because they tend to be repetitive connection failure messages.  As of socket.io-redis 3.0.0, those repetitive failure messages are being forwarded through the adapter as well.  But we still want to support socket.io-redis < 3, so we need to catch and handle errors at the pub/sub client level, plus our way is better because it doesn't fill up the logs.